Major update: - Convert to async for faster performance - Use postgresql since SQLite does not have async functionality
