+++ app.py
@@ -1,0 +1,3 @@
+from service import create_app
+
+app = create_app()
