--- app/models/base.py
@@ -1,4 +1,0 @@
-from sqlmodel import create_engine, SQLModel
-import config
-
-engine = create_engine(config.sqlite_path, echo=True)
