+++ service/models/__init__.py
@@ -1,2 +1,1 @@
 from .transaction import Transaction
-from .base import SQLModel, engine
