+++ service/sync/transaction.py
@@ -1,18 +1,25 @@
-from ..models import engine, Transaction
-from sqlmodel import Session, select
+from sqlmodel.ext.asyncio.session import AsyncSession
+from ..models import Transaction
 from ..clients import Bitcoin
+from sqlmodel import select
 from . import utils
+from .. import db
 import config
 
-def sync_transactions():
+async def sync_transactions():
     utils.log_message("Syncing transactions")
     client = Bitcoin(config.endpoint, rid=config.rid)
 
-    with Session(engine) as session:
+    async with AsyncSession(db.async_engine) as session:
         statement = select(Transaction).where(Transaction.closed)
-        results = session.exec(statement)
+        results = await session.exec(statement)
 
         for transaction in results:
+            data = client.make_request("sendrawtransaction", [transaction.raw_tx])
+
+            if data["result"]:
+                continue
+
             data = client.make_request("decoderawtransaction", [transaction.raw_tx])['result']
 
             input_txid = data["vin"][0]["txid"]
@@ -23,5 +30,5 @@ def sync_transactions():
             if data["result"]:
                 transaction.closed = False
                 session.add(transaction)
-                session.commit()
+                await session.commit()
