+++ service/utils.py
@@ -1,0 +1,4 @@
+from uuid import uuid4
+
+def get_uuid():
+    return str(uuid4())
